refactor: PieceFixture에 Piece 추가
