+++ src/test/java/chess/model/board/PositionFixture.java
@@ -1,0 +1,11 @@
+package chess.model.board;
+
+import chess.model.position.File;
+import chess.model.position.Position;
+import chess.model.position.Rank;
+
+public class PositionFixture {
+
+    public static final Position A1 = new Position(File.A, Rank.FIRST);
+    public static final Position B2 = new Position(File.B, Rank.SECOND);
+}
