+++ src/test/java/chess/model/piece/PieceFixture.java
@@ -1,0 +1,11 @@
+package chess.model.piece;
+
+public class PieceFixture {
+
+    public static final Piece WHITE_BISHOP = new Bishop(PieceColor.WHITE);
+    public static final Piece BLACK_KNIGHT = new Knight(PieceColor.BLACK);
+    public static final Piece WHITE_PAWN = new Pawn(PieceColor.WHITE);
+    public static final Piece WHITE_KING = new King(PieceColor.WHITE);
+    public static final Piece BLACK_QUEEN = new Queen(PieceColor.BLACK);
+    public static final Piece BLACK_ROOK = new Rook(PieceColor.BLACK);
+}
