+++ src/test/java/chess/model/piece/QueenTest.java
@@ -1,5 +1,6 @@
 package chess.model.piece;
 
+import static chess.model.piece.PieceFixture.BLACK_QUEEN;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.model.position.Distance;
@@ -8,7 +9,6 @@ import org.junit.jupiter.params.provider.CsvSource;
 
 class QueenTest {
 
-    private static final Queen WHITE_QUEEN = new Queen(PieceColor.BLACK);
 
     @ParameterizedTest(name = "QUEEN이 ({0}, {1})로 이동할 때 movable 결과가 {2}로 나온다.")
     @CsvSource({"0,3,true", "3,0,true", "-4,-4,true", "6,-6,true", "3,3,true", "-3,3,true"})
@@ -18,11 +18,10 @@ class QueenTest {
             final boolean result
     ) {
         // given
-        final Queen queen = WHITE_QUEEN;
         final Distance distance = new Distance(rank, file);
 
         // when
-        final boolean movable = queen.movable(distance);
+        final boolean movable = BLACK_QUEEN.movable(distance);
 
         // then
         assertThat(movable).isEqualTo(result);
